ability to transfer funds

+++ src/components/users/NewAcc.js
@@ -8,13 +8,13 @@ export default function NewAcc({api, setFunc, setShow, setLogin}){
     const password = useRef("");
     const firstname = useRef("");
     const lastname = useRef("");
+    const [msg, setMsg] = useState('');
     const [isLoading, setLoading] = useState(false);
 
 
     const submit = () => {
       setLoading(true);
-      setShow(false);
-      setLogin(true);
+      
 
       const body = {
         username: username.current,
@@ -24,16 +24,26 @@ export default function NewAcc({api, setFunc, setShow, setLogin}){
       };
 
       axios.post(`${api}/addUser`, body).then(res => {
-        localStorage.setItem('user', JSON.stringify(res.data.user));
-        localStorage.setItem("apiKey", res.data.apiKey);
-        setLoading(false);
-        window.location.reload();
+        if (res.data.status === 'Someone with that username already exists!'){
+          setLoading(false);
+          setMsg(res.data.status);
+        }
+        else {
+          setShow(false);
+          setLogin(true);
+          localStorage.setItem('user', JSON.stringify(res.data.user));
+          localStorage.setItem("apiKey", res.data.apiKey);
+          setLoading(false);
+          window.location.reload();
+        }
+        
       })
       .catch(error => console.log(error));
     }
 
     return(
       <div className="form-container">
+        <div style={{color: 'red', display: msg ? 'block' : 'none'}}>{msg}</div>
         First Name: <input type='text' onChange={(e) => firstname.current = e.target.value}></input><br/>
         Last Name: <input type='text' onChange={(e) => lastname.current = e.target.value}></input><br/>
         Username: <input type='text' onChange={(e) => username.current = e.target.value}></input><br/>
